requirements.txt
ccxt
pandas
schedule
python-telegram-bot==13.15
